% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release-summary.R
\name{get_issues_summary}
\alias{get_issues_summary}
\title{Get a summary on issues of the project.}
\usage{
get_issues_summary(repo, from, until)
}
\arguments{
\item{repo}{The repository name.}

\item{from}{The start date of pull request closed time that the summary operate on.}

\item{until}{The end date of pull request closed time that the summary operate on.}
}
\description{
Get a summary on issues of the project.
}
\note{
The API is experimental and subject to change.
}
\examples{
\donttest{

repo <- "kubeflow/common"
from <- "2020-05-10"
until <- "2020-05-19"
get_issues_summary(repo, from, until)
}

}
