% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release-summary.R
\name{get_release_summary}
\alias{get_release_summary}
\title{Get a summary that can be used for release notes of the project.}
\usage{
get_release_summary(repo, from, until, labels_mapping = NULL)
}
\arguments{
\item{repo}{The repository name.}

\item{from}{The start date of pull request or issue closed time that the summary operate on.}

\item{until}{The end date of pull request or issue closed time that the summary operate on.}

\item{labels_mapping}{The list of mappings from labels to section text in the summary.}
}
\description{
Get a summary that can be used for release notes of the project.
}
\note{
The API is experimental and subject to change.
}
\examples{
\donttest{

repo <- "kubeflow/common"
from <- "2020-05-10"
until <- "2020-05-19"
labels_mapping <- list(
  "size/S" = "Small Changes",
  "size/L" = "Large Changes",
  "size/M" = "Medium Changes")
get_release_summary(repo, from, until, labels_mapping)
}

}
