% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_watchers}
\alias{get_watchers}
\title{Get a list of watchers of a repo.}
\usage{
get_watchers(repo, info_type = "login")
}
\arguments{
\item{repo}{The repository name.}

\item{info_type}{The type of information we get from user profile, e.g. "login" or "email".}
}
\description{
Get a list of watchers of a repo.
}
\examples{
\donttest{

repo <- "terrytangyuan/maintainer-tools"
get_watchers(repo)
}

}
